Apply Copy-Reference(CR) Method to Update JAGS MCMC outputs under MAR for Generalized Linear Model

Internal function to obtain Copy-Reference(CR) MCMC from an MAR object.

glm_MI_CR(object, treatment, start = NULL, end = NULL, thin = NULL,
  exclude_chains = NULL, subset = FALSE, seed = 5432, mess = FALSE,
  ...)

Arguments

object

an object of class remiod

treatment

the variable name of treatment. Reference level of treatment should be coded as 0.

start

first iteration to be used

end

last iteration to be used

thin

thinning to be applied

exclude_chains

optional vector of numbers, indexing MCMC chains to be excluded from the output

subset

subset of parameters (columns of the mcmc object) to be used

seed

optional seed value.

mess

logical, should messages be displayed?

...

optional arguments pass from main function.

Value

A matrix of MCMC samples with all monitored parameters.A subset of the MCMC sample can be selected using start, end and thin.