Listing the sequence of models used for imputation

Listing the sequence of models used for imputation

list.models(object, details = FALSE, print = TRUE)

Arguments

object

an object of class remiod

details

logical. Default is FALSE, where listing all models in formula format. If TRUE, details of each models will be presented.

print

logical. Default is TRUE to print all imputation models or detailed imputation models.

Value

a list of formula of imputation models. If details=TRUE, information on the conditional distributions of the covariates in each imputation models. Note: the sequence of conditional models together specifies the joint distribution.

Examples

# \donttest{
# data(schizow)

test = remiod(formula = y6 ~ tx + y0 + y1 + y3, data = schizow,
              trtvar = 'tx', algorithm = 'jags', method="MAR",
              ord_cov_dummy = FALSE, n.adapt = 10, n.chains = 1,
              n.iter = 10, thin = 2, warn = FALSE, seed = 1234)
#> NOTE: Stopping adaptation
#> 
#> 

list.models(test)
#>   Order model_formula         
#> 1 1     y0 ~ tx               
#> 2 2     y1 ~ tx + y0          
#> 3 3     y3 ~ tx + y0 + y1     
#> 4 4     y6 ~ tx + y0 + y1 + y3
# }