Summarize the results from an object of class remiod

Obtain and print the summary, (fixed effects) coefficients (coef) and credible interval (confint).

summary(object, ...)

# S3 method for remiod
summary(object, start = NULL, end = NULL, thin = NULL,
  quantiles = c(0.025, 0.975), outcome = NULL, exclude_chains = NULL,
  warn = TRUE, mess = TRUE, ...)

# S3 method for summary.remiod
print(x, digits = 3, ...)

# S3 method for summary.remiod
coef(object, start = NULL, end = NULL,
  thin = NULL, subset = NULL, exclude_chains = NULL, warn = TRUE,
  mess = TRUE, ...)

Arguments

object

object inheriting from class 'remoid'

...

additional, optional arguments

trunc

named list specifying limits of truncation for the distribution of the named incomplete variables (see the vignette ModelSpecification)

hyperpars

list of hyper-parameters, as obtained by default_hyperpars()

scale_vars

named vector of (continuous) variables that will be centred and scaled (such that mean = 0 and sd = 1) when they enter a linear predictor to improve convergence of the MCMC sampling. Default is that all numeric variables and integer variables with >20 different values will be scaled. If set to FALSE no scaling will be done.

custom

named list of JAGS model chunks (character strings) that replace the model for the given variable.

append_data_list

list that will be appended to the list containing the data that is passed to rjags (data_list). This may be necessary if additional data / variables are needed for custom (covariate) models.

progress.bar

character string specifying the type of progress bar. Possible values are "text" (default), "gui", and "none" (see update). Note: when sampling is performed in parallel it is not possible to display a progress bar.

quiet

logical; if TRUE then messages generated by rjags during compilation as well as the progress bar for the adaptive phase will be suppressed, (see jags.model)

keep_scaled_mcmc

should the "original" MCMC sample (i.e., the scaled version returned by coda.samples()) be kept? (The MCMC sample that is re-scaled to the scale of the data is always kept.)

modelname

character string specifying the name of the model file (including the ending, either .R or .txt). If unspecified a random name will be generated.

modeldir

directory containing the model file or directory in which the model file should be written. If unspecified a temporary directory will be created.

overwrite

logical; whether an existing model file with the specified <modeldir>/<modelname> should be overwritten. If set to FALSE and a model already exists, that model will be used. If unspecified (NULL) and a file exists, the user is asked for input on how to proceed.

keep_model

logical; whether the created JAGS model file should be saved or removed from (FALSE; default) when the sampling has finished.

start

the first iteration of interest (see window.mcmc)

end

the last iteration of interest (see window.mcmc)

thin

thinning interval (integer; see window.mcmc). For example, thin = 1 (default) will keep the MCMC samples from all iterations; thin = 5 would only keep every 5th iteration.

quantiles

posterior quantiles

outcome

specify outcome variable to select imputation model(s) to summarize. Default generates summaries for all models.

exclude_chains

optional vector of the index numbers of chains that should be excluded

warn

logical; should warnings be given? Default is TRUE.

mess

logical; should messages be given? Default is TRUE.

x

an object of class summary.remiod

digits

the minimum number of significant digits to be printed in values.

subset

subset of parameters/variables/nodes (columns in the MCMC sample). Follows the same principle as the argument monitor_params and selected_parms.

Value

summary information, including parameter posterior mean, posterior SD, quantiles, tail probability tail-prob, Gelman-Rubin criterion GR-crit, the ratio of the Monte Carlo error and posterior standard deviation) for specified parameters MCE/SD.

Examples

# \donttest{
# data(schizow)

test = remiod(formula = y6 ~ tx + y0 + y1 + y3, data = schizow,
              trtvar = 'tx', algorithm = 'jags', method="MAR",
              ord_cov_dummy = FALSE, n.adapt = 50, n.chains = 1,
              n.iter = 50, thin = 2, warn = FALSE, seed = 1234)
#> NOTE: Stopping adaptation
#> 
#> 

summary(object = test, outcome = c("y6","y3"))
#> 
#> 
#> Call:
#> clm_imp_custom(formula = y6 ~ tx + y0 + y1 + y3, data = data, 
#>     trtvar = trtvar, n.chains = n.chains, n.adapt = n.adapt, 
#>     n.iter = n.iter, thin = thin, monitor_params = list(imps = TRUE, 
#>         other_models = TRUE, other = c(eta_monits)), refcats = refcats, 
#>     model_order = model_order, seed = seed, warn = warn, mess = mess, 
#>     ord_cov_dummy = ord_cov_dummy)
#> 
#> 
#> # --------------------------------------------------------------------- #
#>   Bayesian cumulative logit model for "y6"
#> # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - #
#> 
#> Posterior summary:
#>               Mean     SD   2.5%  97.5% tail-prob. GR-crit MCE/SD
#> beta_y6_tx  -1.407 0.2389 -1.879 -1.022          0               
#> beta_y6_y0  -0.341 0.1000 -0.503 -0.145          0               
#> beta_y6_y1   0.578 0.0895  0.435  0.767          0               
#> beta_y6_y3   1.478 0.0817  1.329  1.621          0               
#> gamma_y6[1] -1.792 0.2402 -2.232 -1.422          0               
#> gamma_y6[2] -4.668 0.2673 -5.123 -4.173          0               
#> gamma_y6[3] -6.070 0.3441 -6.603 -5.464          0               
#> 
#> 
#> # --------------------------------------------------------------------- #
#>   Bayesian cumulative logit model for "y3"
#> # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - #
#> 
#> Posterior summary:
#>                Mean     SD   2.5%  97.5% tail-prob. GR-crit MCE/SD
#> beta_y3_tx  -1.2169 0.2007 -1.670 -0.929       0.00               
#> beta_y3_y0   0.0965 0.0827 -0.012  0.266       0.24               
#> beta_y3_y1   1.5373 0.1372  1.342  1.773       0.00               
#> gamma_y3[1] -1.3473 0.2092 -1.685 -0.891       0.00               
#> gamma_y3[2] -3.8126 0.3566 -4.464 -3.136       0.00               
#> gamma_y3[3] -5.8234 0.3733 -6.392 -5.226       0.00               
#> 
#> 
#> # ----------------------------------------------------------- #
#> 
#> MCMC settings:
#> Iterations = 52:100
#> Sample size per chain = 25 
#> Thinning interval = 2 
#> Number of chains = 1 
#> 
#> Number of observations: 437 
# }