Apply a MI method following initial run of remoid function

Apply a MI method following initial run of remoid function

Usage

updateMI(object, method = c("MAR", "J2R", "CR", "delta"), delta = 0,
  mi.setting = NULL, mess = TRUE)

Arguments

object

object inheriting from class remoid.

method

a method for obtaining multiple-imputed dataset. Options include MAR, J2R, CR, and delta adjustment. Default is MAR.

delta

specific value used for Delta adjustment, applicable only for method="delta".

mi.setting

a list of arguments for extracting MI data set, which will be used to update the one in remoid object. Default is NULL, meaning no update to the mi.setting in remoid object.

mess

logical; should messages be given? Default is TRUE.